[librpfile] RecursiveScan_win32.cpp: Use d_type.
This lets us change the std::forward_list<> to return uint8_t again.

TODO: Add a "file attributes to d_type" function, and update
FileSystem::get_file_d_type() to use it.

+++ src/gtk/config/CacheCleaner.cpp
@@ -302,7 +302,7 @@ rp_cache_cleaner_run(RpCacheCleaner *cleaner)
 	// Recursively scan the cache directory.
 	// TODO: Do we really want to store everything in a list? (Wastes memory.)
 	// Maybe do a simple counting scan first, then delete.
-	forward_list<pair<string, uint32_t> > rlist;
+	forward_list<pair<string, uint8_t> > rlist;
 	int ret = recursiveScan(cacheDir.c_str(), rlist);
 	if (ret != 0) {
 		// Non-image file found.
